Fix daily 'boolean value of NA is ambiguous' + Deep Dive history selector & accuracy
Root cause: newer yfinance/pandas return OHLCV as nullable extension dtypes
(Float64/Int64); _normalise preserved them, so .values was an ExtensionArray and
the detectors raised 'NA is ambiguous' on the daily (un-aggregated) frame
(weekly/monthly worked because aggregation forced numpy floats). Fix: coerce to
numpy float64 in _normalise. Regression test added.

Deep Dive also gains:
- a Chart history selector (1Y/3Y/5Y/10Y/Max) over a generous fetch, so you see
  full history instead of just the last 300 bars,
- an Accuracy & historical-edge section (historical hit-rate, out-of-sample,
  occurrences, avg forward return, verdict) — the edge walk-forward is bounded
  to a solid sample to stay fast.

148 tests pass.

+++ nsetrade/data/base.py
@@ -65,7 +65,10 @@ class DataProvider(ABC):
             raise ValueError(f"data is missing required columns: {missing}")
         df = df[OHLCV_COLUMNS].copy()
         for col in OHLCV_COLUMNS:
-            df[col] = pd.to_numeric(df[col], errors="coerce")
+            # force NumPy float64 — pandas nullable dtypes (Float64/Int64) keep a
+            # masked array whose `.values` is an ExtensionArray, which makes the
+            # detectors raise "boolean value of NA is ambiguous".
+            df[col] = pd.to_numeric(df[col], errors="coerce").astype("float64")
         df = df.dropna(subset=["open", "high", "low", "close"])
         df = df[~df.index.duplicated(keep="last")].sort_index()
         if isinstance(df.index, pd.DatetimeIndex) and df.index.tz is not None:
